Refactor: move update code from project_comms to fmt_update
