+++ firmware/fmt_rx.in.c
@@ -2,6 +2,7 @@
  * Generated by gen-firment.py based on .proto file.
  * Provides first-level handling of received messages.
  * */
+#include <fmt_update.h>
 #include <fmt_rx.h>
 #include <fmt_spi.h>       // fmt_getMsg()
 #include <fmt_log.h>       // fmt_sendLog()
