+++ firmware/fmt_update.in.h
@@ -3,6 +3,9 @@
  * Provides address where new images should be written (so BL will see them).
  *   This comes from CMake, as it is also needed by bootloader.
  */
+
+#include "messages.pb.h"
+#include <stdint.h>
  
 // This needs to be gotten from project-level cmake (shared with bootloader). 
 // XMC_FLASH_SECTOR_(0-8): Bootloader, 256kB total. 
@@ -21,3 +24,6 @@
 
 #define CHUNKS_PER_PAGE_MAX  (FLASH_PAGE_SIZE / IMAGE_CHUNK_MAX_SIZE)
 #define PAGES_PER_SECTOR_MAX (IMAGE_RX_SECTOR_SIZE / FLASH_PAGE_SIZE)
+
+#define USE_ImageData
+void handleImageData(ImageData msg);
